сделал ScrollToTop при смене предложения при нажатии в NearPlaces
