+++ src/pages/offer-page/offer-page.tsx
@@ -9,6 +9,7 @@ import { loadDetailOffer, loadOfferNearOffers, loadOfferReviews } from '../../st
 import { fetchDetailOfferAction, fetchOfferNearOffersAction, fetchOfferReviewsAction } from '../../store/api-actions';
 import { OfferId } from '../../types/offer';
 import { APP_TITLE, EMPTY_DETAIL_OFFER } from '../../const';
+import ScrollToTop from '../../hooks/scroll-to-top';
 
 function OfferPage(): JSX.Element {
   const params = useParams();
@@ -34,6 +35,7 @@ function OfferPage(): JSX.Element {
 
   return (
     <div className="page">
+      <ScrollToTop />
       <Helmet>
         <title>{`${APP_TITLE}: offer`}</title>
       </Helmet>
